requirements.txt
azure-functions
prefect >= 2.1.1